ESCAPE key handling on modals + new option for Assist Conv + bug fixes
Fixed an issue with app icon not being resized in About window.
Added "are you sure you want to leave the window" when pressing ESC key.
Fixed (this this, it's fine, but visually ugly) overlapping items in BodySlide Filters Editor.
Fixed a bug when exiting the settings window: no message where displayed if the user was about to lose the value of the checkForUpdate.
Added an option to target the Assisted Conversion only on the "meshes" directory.
Fixed the update checker at startup that was not displaying "update available" in the menu bar.